refactor(engine): consolidate foundational primitives into lib/src/engine with formal attribution
- Consolidate Tim Lehmann's foundational rendering primitives into lib/src/engine/
- Add formal per-directory attribution (lib/src/engine/ATTRIBUTION.md, LICENSE, THIRD_PARTY_NOTICES)
- Keep custom scopes in lib/src/renderer/ and re-export via internal barrel
- Update import paths across affected widgets and test suites
- Update CHANGELOG.md with engine organization chore entry
