+++ README.md
@@ -998,7 +998,7 @@ Focused, self-contained demos — one widget, one file, runnable standalone:
 ## Documentation
 
 - **[Migration Guide (0.x to 1.0.0)](docs/MIGRATION_0.x_TO_1.0.md)** — Step-by-step upgrade guide for 1.0.0 breaking changes.
-- **[Architecture & Guidelines](ARCHITECTURE.md)** — Core design principles and internal architecture.
+- **[Architecture & Guidelines](docs/ARCHITECTURE.md)** — Core design principles and internal architecture.
 - **[Platform Support](docs/PLATFORM_SUPPORT.md)** — Platform matrices and rendering pipeline compatibility.
 
 
@@ -1012,9 +1012,13 @@ Contributions are welcome. For major changes, open an issue first to discuss you
 MIT — see the [LICENSE](LICENSE) file for details.
 
 
-## Credits
+## Acknowledgments & Credits
 
-**Special thanks** to the [whynotmake-it](https://github.com/whynotmake-it) team for their [`liquid_glass_renderer`](https://github.com/whynotmake-it/flutter_liquid_glass/tree/main/packages/liquid_glass_renderer) (MIT), whose shader pipeline, texture capture, and chromatic aberration work forms the foundation of the rendering engine in this library.
+Special thanks and sincere recognition to **Tim Lehmann** ([whynotmake.it](https://github.com/whynotmake-it)) and the [whynotmake-it](https://github.com/whynotmake-it/flutter_liquid_glass) team:
+- [`liquid_glass_renderer`](https://github.com/whynotmake-it/flutter_liquid_glass/tree/main/packages/liquid_glass_renderer) (MIT License): Pioneered the foundational fragment shader pipeline, blend group tracking, and texture capture architecture. Foundational work is maintained in-tree in [`lib/src/engine/`](lib/src/engine/) with full attribution.
+- [`motor`](https://github.com/whynotmake-it/rivership/tree/main/packages/motor) (MIT License): Designed the `CupertinoMotion` spring physics model and presets adapted with Flutter physics primitives in [`lib/utils/glass_spring.dart`](lib/utils/glass_spring.dart).
+
+For complete licenses, copyright notices, and local modification records, see [THIRD_PARTY_NOTICES](THIRD_PARTY_NOTICES) and [`lib/src/engine/ATTRIBUTION.md`](lib/src/engine/ATTRIBUTION.md).
 
 ## Links
