+++ example/integration_test/press_scripted_test.dart
@@ -13,7 +13,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 
 import 'package:liquid_glass_widgets/liquid_glass_widgets.dart';
-import 'package:liquid_glass_widgets/src/renderer/stretch.dart';
+import 'package:liquid_glass_widgets/src/engine/stretch.dart';
 import 'package:liquid_glass_widgets_example/constants/glass_settings.dart';
 
 void main() {
