+++ lib/src/engine/liquid_glass_render_scope.dart
@@ -1,7 +1,15 @@
+// Copyright 2024-2025 Tim Lehmann for whynotmake.it
+//
+// SPDX-License-Identifier: MIT
+//
+// Originally from liquid_glass_renderer (whynotmake.it).
+// Maintained and evolved in-tree for liquid_glass_widgets.
+// See lib/src/engine/ATTRIBUTION.md for provenance and modification history.
+
 // ignore_for_file: public_member_api_docs
 
 import 'package:flutter/widgets.dart';
-import 'liquid_glass_renderer.dart';
+import 'liquid_glass_settings.dart';
 
 class LiquidGlassRenderScope extends InheritedWidget {
   /// Creates a new [LiquidGlassRenderScope].
