+++ lib/src/engine/liquid_shape.dart
@@ -1,7 +1,14 @@
+// Copyright 2024-2025 Tim Lehmann for whynotmake.it
+//
+// SPDX-License-Identifier: MIT
+//
+// Originally from liquid_glass_renderer (whynotmake.it).
+// Maintained and evolved in-tree for liquid_glass_widgets.
+// See lib/src/engine/ATTRIBUTION.md for provenance and modification history.
+
 import 'dart:math' as math;
 
 import 'package:flutter/widgets.dart';
-import 'liquid_glass_renderer.dart';
 
 /// Represents a shape that can be used by a [LiquidGlass] widget.
 // ignore: deprecated_member_use
