+++ lib/src/renderer/fragment_shader_extensions.dart
