+++ lib/types/interaction_notification.dart
