+++ test/renderer/glass_glow_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:liquid_glass_widgets/src/renderer/glass_glow.dart';
+import 'package:liquid_glass_widgets/src/engine/glass_glow.dart';
 
 void main() {
   // ──────────────────────────────────────────────────────────────────────────
