+++ test/renderer/liquid_glass_test.dart
@@ -8,7 +8,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:liquid_glass_widgets/liquid_glass_widgets.dart';
-import 'package:liquid_glass_widgets/src/renderer/liquid_glass.dart';
+import 'package:liquid_glass_widgets/src/engine/liquid_glass.dart';
 
 import '../shared/test_helpers.dart';
