+++ test/renderer/liquid_shape_test.dart
@@ -3,7 +3,7 @@ import 'dart:ui' as ui;
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:liquid_glass_widgets/liquid_glass_widgets.dart';
-import 'package:liquid_glass_widgets/src/renderer/liquid_glass.dart';
+import 'package:liquid_glass_widgets/src/engine/liquid_glass.dart';
 
 void main() {
   // ──────────────────────────────────────────────────────────────────────────
