+++ test/renderer/stretch_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:liquid_glass_widgets/src/renderer/internal/liquid_glass_self_scale_scope.dart';
-import 'package:liquid_glass_widgets/src/renderer/stretch.dart';
+import 'package:liquid_glass_widgets/src/renderer/liquid_glass_self_scale_scope.dart';
+import 'package:liquid_glass_widgets/src/engine/stretch.dart';
 
 void main() {
   // ──────────────────────────────────────────────────────────────────────────
